requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
flask-sqlalchemy==2.5.1
psycopg2==2.9.2
